Fix mutation algorithm, add test
Previously, we were calculating the index of the next mutation as
mutation_position.round(). This caused index 0 to be half as likely
to mutate as the other bytes, because mutation_position is non-negative:

- If mutation_position was in [0.0, 0.5), it mapped to index 0.
- If it was in [0.5, 1.5), it mapped to index 1.
- If it was in [1.5, 2.5), it mapped to index 2, and so on.

In the first example, index 0 gets short-changed: its range is half
the size of the other indexes.

This commit changes the calculation to mutation_position.floor(),
which should be fair to all the indexes. Now a mutation_position in
the range [0.0, 1.0) will map to index 0.