Allow Population to grow over time
Previously, Plasma kept its Population at a constant size. When this size
was small, it would lead to inbreeding: there would be little variation in
plasmas, and sometimes the system would get stuck on a single plasma. But
when the Population size was larger, the user's input would have less
effect, because the user's chosen plasmas would be greatly outnumbered by
the existing plasmas.

This commit starts Plasma off with a smaller Population, but allows that
Population to grow over time. This gives the user more control at the
beginning, while hopefully preventing inbreeding later on.