Replace mul_add() with *, +
I was using fused multiply-add in a couple of places, thinking that it
might be faster than using regular multiplication and addition. However,
mul_add() got slower with Rust 1.12: it was causing the first few frames
to take longer than normal to render. Getting rid of mul_add() brought the
render speed back to normal.