Replace from_hsv() with gamma-correct from_hsl()
Randomly-chosen HSV parameters have a bias toward darker colors. If these
parameters are bytes, there are 65536 ways to get black, but only 256 to
get white. This results in plasmas that tend to be on the dark side. HSL
doesn't have this particular problem, because the color model has symmetry
between black (anything with L = 0.0) and white (anything with L = 1.0).

The tradeoff is that the number of distinct colors we can represent in 3
bytes is slightly smaller, but this is a relatively minor cost.

Also in this commit: interpolation inside from_hsl() is now gamma correct,
which was not the case in the previous HSV implementation.