Use square-shaped HSL for ControlPoint colors
Previously, ControlPoint Genes stored colors using a cylindrical
coordinate system (HSL). This commit modifies HSL to use Cartesian
coordinates instead.

Why? Because when Genes mutate, their bytes do not wrap around: the byte
value 255 cannot increase to the value 0. In terms of hue, this means that
an angle of 359 (slightly-purplish red) cannot mutate to 0 (pure red),
even though the two colors are visually close to each other. Cartesian
coordinates handle this particular situation much better.

If I had to do this over, it would have been easier to start with a
different color space, such as YCbCr or YCgCo. But square-shaped HSL
does come with a nice benefit: if you use it to generate random colors,
and your random inputs are uniformly distributed, then all hues will
be equally represented in the output.