Add palette optimizer to ColorMapper
Previously, ColorMapper's lookup table was also its palette, which meant
that the sizes of the two had to be the same. This commit separates
palette generation into its own routine, so that in the future they can
be changed independently.

Palette generation currently uses the k-means algorithm to pick an
adaptive palette. This is slow, mainly because it uses a lot of floating-
point math. This will be fixed later. As for now, I'm happy to see that it
works!