Flush out command-line flags, unify param structs
This commit adds a bunch of command-line flags. At the moment, only a
few of them are hooked up to anything, but they do give a sense of what
functionality the final program should have.

This change also introduces a single PlasmaSettings struct, which replaces
the two separate structs that existed before (one for interactive use, and
one for dumping to a file).