Speed up test_asyncrenderer_cancellation() a little
This test is nondeterministic: it has to wait until an image is rendered,
and it cannot check on the progress of the render, so it has to sleep for
a few milliseconds instead.

Previously, it was sleeping for hundreds of milliseconds: AsyncRenderer
can take ~70ms for the first render, so the test slept for 200ms to be on
the safe side. This commit adds code to warm up the renderer beforehand,
so that the test only has to blindly sleep for 10ms. I could probably
reduce it further, but I like being on the safe side.

This commit also simplifies the logic in the second part of the test,
where the test has to tell image A from image B.