Actually optimize GIFs
Previously, when optimizing GIFs by using transparent pixels, the code
made as many pixels transparent as possible. This often had the effect
of *increasing* file sizes, because it broke long runs of a single color
into many shorter runs.

This commit makes the optimizer more judicious. Before it makes a run of
pixels transparent, it estimates whether doing so would make the image
more lengthy to encode (using RLE as a heuristic). If so, it leaves those
pixels opaque.