Adjust formulas to be biased toward slow plasmas
Using the --loop-duration flag to make a shorter plasma causes the whole
animation to be sped up. The default loop duration is 60 seconds: changing
it to 10 seconds would cause the animation to play at 6x speed. If you
want a short loop time, you want to find a plasma that's really slow.

This commit makes finding slow plasmas easier. It does this by adjusting
to_ifloat() to be biased toward small numbers: to_ifloat() is used for
animation parameters, so this results in a bias toward slower animations.

The overall range of to_ifloat() is unchanged, so all plasmas that were
possible before are still possible -- but they will have different genomes
than before.