Revise color quantization to be gamma-correct
This commit updates ColorMapper to use LinearColors when calculating
the optimal palette. Its interface is unchanged -- this commit only
changes the internal calculations.

Previously, ColorMapper used gamma-encoded values to calculate
color distances. I did this both because it was faster, and because
I thought that gamma-encoded distance corresponded with perceived
distance. But I was wrong! I was confusing perceived-amount-of-change
with ability-to-perceive-change.

I dreaded fixing this. But now that I have LinearColor, it's easy to
go back and make this code both correct *and* fast. Yay!