Move Color, LinearColor to their own submodule
I was having trouble remembering which structs were defined where. Time to
reorganize!

Color and LinearColor used to be in gradient.rs, because the gradient code
was the first to need them. But as the program has grown, it makes more
sense to give them their own place. I'm giving them a submodule because
there's other color-related code (e.g., palette manipulation) that I would
like to move to the same place.