Improve stderr output for output errors
Previously, if there was an error writing output to a file (e.g., passing
a directory name to the -o option instead of a file), the plasma app would
panic. Now, the panic message would contain the necessary details, but it
gave the impression of something being wrong with the plasma itself.

Now the plasma prints out a nice, explanatory error message.

a filesystem error (e.g., -o directory instead of -o file)