Increase palette range when dithering is enabled
When doing nearest-color quantization, we want a palette that minimizes
(target color) - (nearest palette color). This is what Palette::new() has
been doing so far, and by this measure, it was doing a pretty good job.

However, when dithering is enabled, we care about more than just the
distance to the nearest palette color -- we also care about whether our
target color is within the palette's gamut. In other words, can we get
arbitrarily close to our target color by mixing palette colors together?

This commit tweaks Palette::new() to maximize the palette's range when
dithering is enabled. It's a little hacky, but it makes the dithered
plasmas look a lot closer to the corresponding full-color ones.